+++ soluciones/solucion_ufds.cpp
@@ -13,7 +13,7 @@
 #include <cmath>
 using namespace std;
 
-const int EPS = 1e-7;
+const double EPS = 1e-7;
 const int ID_FRAGIL = 0;    // ID del nenúfar frágil
 
 struct tNenufar {
